requirements.txt
python-telegram-bot==22.3
Requests==2.32.4
